Compatibility with python3 before 3.5 (PEP 0448)
read_passes currently requires Python 3.5 or later, due to
its use of "**" for dictionary unpacking within a "dictionary
display".  The generalization of the "**" syntax was added
in Python 3.5 via:
  https://www.python.org/dev/peps/pep-0448/

This patch avoids this syntax by using dict.update, and adds a docstring
to the function (to ensure that I got the meaning correct).